refactor: ProfileCard와 MusicListItem을 entity 레이어로 이동
