+++ app/msw/handlers.ts
@@ -1,0 +1,3 @@
+import { dormitoryHandlers } from "@/entities/dormitory/api/dormitory.handlers";
+
+export const handlers = [...dormitoryHandlers];
