+++ src/features/wake-up-music/ui/WakeUpMusicSection.tsx
@@ -3,7 +3,7 @@
 import { ReactNode, useState } from "react";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import Music from "@/shared/asset/svg/Music";
-import { MusicListItem } from "@/shared/ui/MusicListItem";
+import { MusicListItem } from "@/entities/music/ui/MusicListItem";
 import { Calendar } from "@/shared/ui/Calendar";
 import { TextButton } from "@/shared/ui/Button/TextButton";
 import TextField from "@/shared/ui/textField";
