--- src/shared/api/msw/handlers.ts
@@ -1,3 +1,0 @@
-import { dormitoryHandlers } from '@/entities/dormitory/api/dormitory.handlers';
-
-export const handlers = [...dormitoryHandlers];
