--- src/shared/api/msw/worker.ts
@@ -1,4 +1,0 @@
-import { setupWorker } from 'msw/browser';
-import { handlers } from './handlers';
-
-export const worker = setupWorker(...handlers);
